+++ NEWS.md
@@ -1,6 +1,7 @@
 # mapSpain 0.1.1
 
 * Fix CRAN submission.
+* Added esp_get_grid_prov() and esp_get_grid_ccaa().